9월19일 45번까지

+++ codeup/39.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 class Main{
   public static void main(String[] args){
     Scanner sc = new Scanner(System.in);
-    int num1 = sc.nextInt();
-    int num2 = sc.nextInt();
-    int sum = num1 + num2;
+    long num1 = sc.nextLong();
+    long num2 = sc.nextLong();
+    long sum = num1 + num2;
     System.out.printf("%d", sum);
   }
 }
